% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_test.R
\name{acc_test}
\alias{acc_test}
\title{Returns the rsme and median absolute error of different models}
\usage{
acc_test(
  data,
  seed,
  formula1 = "sqrtgr ~ logground",
  formula2 =
    "sqrtgr ~ logground + roofflat*Exposure +\\n                     roofflat*winter_wind +\\n                     log(Size) + temp_avg + Heated + Parapet",
  tree_vars = c(7, 11, 13, 19:22, 25:27, 33, 35),
  rf_y = "gr"
)
}
\arguments{
\item{train_data}{The training data}

\item{test_data}{The testing data}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
