% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_header.R
\name{add_header}
\alias{add_header}
\title{Adds a header row}
\usage{
add_header(
  folder_path,
  existing_colums = c("date", "value"),
  new_columns = c("measurement", "building_code", "mma")
)
}
\arguments{
\item{folder_path}{the path of the folder being edited}

\item{existing_columns}{The names of the existing column}

\item{new_colums}{The names of the new columns}

\item{rightdate}{The correct start date}
}
\description{
This functions adds a header row to the date and names the column.
}
