% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_dates.R
\name{add_missing_dates}
\alias{add_missing_dates}
\title{Adds missing date in data}
\usage{
add_missing_dates(df)
}
\arguments{
\item{df}{The data to add dates to}
}
\description{
Adds missing date in data
}
