% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_var.R
\name{code_var}
\alias{code_var}
\title{sets the code variable based on folder name.}
\usage{
code_var(path, folder, column)
}
\arguments{
\item{path}{the path of the folder containing the subfolder}

\item{folder}{The names of the subfolder}

\item{colum}{The column number to make the update to}
}
\description{
Takes the folder name and sets the code variable based on that
}
