% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_dir_fix.R
\name{date_dir_fix}
\alias{date_dir_fix}
\title{Fixing dates moving the wrong direction}
\usage{
date_dir_fix(data, firstmonth, lastmonth, startdate)
}
\arguments{
\item{data}{the data file to fix date order}

\item{firstmonth}{The number corresponding to the first month on the graph}

\item{lastmonth}{The number corresponding to the last month on the graph}

\item{startdate}{the 1st date on the graph}
}
\description{
Fixing dates moving the wrong direction
}
