% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_fix.R
\name{date_fix}
\alias{date_fix}
\title{November 1 1964 is recorded and nov 26 1964 on the digizer}
\usage{
date_fix(data, enddate, wrongdate = "1964-12-01", rightdate = "1964-11-01")
}
\arguments{
\item{data}{the data file to fix end month}

\item{enddate}{The end date on the graph}

\item{wrongdate}{The inncorrect start date}

\item{rightdate}{The correct start date}
}
\description{
This function exists to fix the issue on the digitizer of recoring the
wrong date.
}
