% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_cv.R
\name{gr_cv}
\alias{gr_cv}
\title{Returns the rsme and median absolute error of different models}
\usage{
gr_cv(
  train_data,
  test_data,
  formula1 = "sqrtgr ~ logground",
  forumula2 =
    "sqrtgr ~ logground + roofflat*Exposure +\\n                  roofflat*winter_wind +\\n                   log(Size) + temp_avg + Heated + Parapet",
  rf_formula = gr ~ ground_max + roofflat + Exposure + winter_wind + Size + temp_avg +
    Heated + Parapet
)
}
\arguments{
\item{train_data}{The training data}

\item{test_data}{The testing data}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
