% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_month_fix.R
\name{last_month_fix}
\alias{last_month_fix}
\title{Fixing dates that end in the wrong month}
\usage{
last_month_fix(
  data,
  firstmonth,
  wrongmonth,
  rightmonth,
  startdate,
  diffyearwrong = TRUE,
  diffyearright = TRUE
)
}
\arguments{
\item{data}{the data file to fix end month}

\item{firstmonth}{The number corresponding to the first month on the graph}

\item{wrongmonth}{The number corresponding to the wrong last month on graph}

\item{rightmonth}{Th number corresponding to the right last month on graph}

\item{startdate}{the 1st date on the graph}

\item{diffyearwrong}{TRUE if the wrong month is in a different year}

\item{diffyearright}{TRUE if the right month is in a different year}
}
\description{
Fixing dates that end in the wrong month
}
