% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_avg_grid.R
\name{temp_avg_grid}
\alias{temp_avg_grid}
\title{Retuns average temp from raster}
\usage{
temp_avg_grid(file, var1 = "t2m")
}
\arguments{
\item{file}{the file to return avg_temp from}

\item{var1}{1st wind variable to use}
}
\description{
Retuns average temp from raster
}
