% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_avg_grid.R
\name{wind_avg_grid}
\alias{wind_avg_grid}
\title{Returns the average wind speed from ERA 5 data}
\usage{
wind_avg_grid(file, var1 = "u10", var2 = "v10")
}
\arguments{
\item{file}{the file to return winter_wind from}

\item{var1}{1st wind variable to use}

\item{var2}{2nd wind variable to use}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
