% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_test.R
\name{acc_test}
\alias{acc_test}
\title{Returns the rsme and median absolute error of different models}
\usage{
acc_test(
  data,
  k = 5,
  formula1 = "sqrtgr ~ logground",
  formula2 = "sqrtgr ~ logground + winter_wind_all",
  rf_formula = gr ~ ground_max + roofflat + Exposure + winter_wind + Size + temp_avg +
    Heated + Parapet
)
}
\arguments{
\item{data}{The full data to perform models on}

\item{formula1}{The 1st formula to use. Default is logground}

\item{formula2}{The 2nd fromula to use Returns as "new model"}

\item{rf_formula}{The formula to use for the random forest}

\item{seed}{The Random number seed used to partition data}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
