% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_dir_fix.R
\name{date_dir_fix}
\alias{date_dir_fix}
\title{Fixing dates moving the wrong direction}
\usage{
date_dir_fix(data, startdate, enddate)
}
\arguments{
\item{data}{the data file to fix date order}

\item{startdate}{the 1st date on the graph}

\item{enddate}{the last date on the graph}
}
\description{
Fixing dates moving the wrong direction
}
