% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_cv.R
\name{gr_cv}
\alias{gr_cv}
\title{Returns the rsme and median absolute error of different models}
\usage{
gr_cv(train_data, test_data)
}
\arguments{
\item{train_data}{The training data}

\item{test_data}{The testing data}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
