% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_cv.R
\name{gr_cv}
\alias{gr_cv}
\title{Returns the rsme and median absolute error of different models}
\usage{
gr_cv(
  train_data,
  test_data,
  formula1 = "sqrtgr ~ logground",
  formula2 = sqrtgr ~ logground + roofflat * Exposure + roofflat * winter_wind +
    log(Size) + temp_avg + Heated + Parapet,
  rf_formula = gr ~ ground_max + roofflat + Exposure + winter_wind + Size + temp_avg +
    Heated + Parapet
)
}
\arguments{
\item{train_data}{The training data}

\item{test_data}{The testing data}

\item{formula1}{The 1st formula to use. Default is logground}

\item{formula2}{The 2nd fromula to use Returns as "new model"}

\item{rf_formula}{The formula to use for the random forest}
}
\description{
Takes train and test data and comuptes different models to compare accuracy
measures
}
