% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_var.R
\name{other_var}
\alias{other_var}
\title{sets variables based on file name}
\usage{
other_var(folder_path, pattern, column, varval)
}
\arguments{
\item{folder_path}{The folder path to get to the files}

\item{pattern}{The pattern for the files chosen}

\item{column}{The column number to make the update to}

\item{varval}{The value to give the variables}
}
\description{
takes files based on pattern and saves all rows of a variable
}
